# Optimise generate of plugin config