--- a/src/parse/asp/config.go
+++ b/src/parse/asp/config.go
@@ -209,9 +209,19 @@ func (i *interpreter) loadPluginConfig(s *scope, pluginState *core.BuildState) {
 	if _, ok := s.config.overlay[key]; ok {
 		return
 	}
+	s.config.overlay[key] = &pyConfig{base: i.pluginConfig(key, pluginState, s.state)}
+}
 
-	cfg := pluginConfig(pluginState, s.state)
-	s.config.overlay[key] = cfg
+func (i *interpreter) pluginConfig(key string, pluginState *core.BuildState, pkgState *core.BuildState) *pyConfigBase {
+	if cfg, wait, first := i.pluginConfigs.GetOrWait(key); cfg != nil {
+		return cfg
+	} else if !first {
+		<-wait
+		return i.pluginConfigs.Get(key)
+	}
+	cfg := &pyConfigBase{dict: pluginConfig(pluginState, pkgState)}
+	i.pluginConfigs.Set(key, cfg)
+	return cfg
 }
 
 func toPyObject(key, val, toType string) pyObject {