Been looking at the [performance graphs](https://please.build/performance.html) which have gone in the bad direction during the v17 work. I can imagine it will be a _bit_ slower but this seems more than I'd expect, and defo wouldn't expect the memory usage to increase the way it has.

Turns out we're spending quite a bit of time & memory in v17 recreating the plugin config (it's redone each time at every `subinclude`, even the global ones). This caches the initial config for each plugin. It also represents the plugin configs as nested config objects rather than normal dicts, so we can take advantage of its base/overlay behaviour to override things.

Before:
```
INFO: Run 1 of 5
INFO: Complete in 7.94s, using 12859852 KB
INFO: Run 2 of 5
INFO: Complete in 7.97s, using 12466732 KB
INFO: Run 3 of 5
INFO: Complete in 7.89s, using 11885564 KB
INFO: Run 4 of 5
INFO: Complete in 7.66s, using 11138540 KB
INFO: Run 5 of 5
INFO: Complete in 7.86s, using 11697292 KB
INFO: Complete, median time: 7.89s, median mem: 11885564.00 KB
```
After:
```
INFO: Run 1 of 5
INFO: Complete in 6.97s, using 9329508 KB
INFO: Run 2 of 5
INFO: Complete in 6.86s, using 9855560 KB
INFO: Run 3 of 5
INFO: Complete in 6.96s, using 9520748 KB
INFO: Run 4 of 5
INFO: Complete in 7.07s, using 9695692 KB
INFO: Run 5 of 5
INFO: Complete in 7.07s, using 10228892 KB
INFO: Complete, median time: 6.97s, median mem: 9695692.00 KB
```
This will do better vs. our normal benchmark (the CPU is faster and has more cores than CircleCI's `xlarge` resource class) but we should see improvement there.